feat: clean增加清理未被版本跟踪的嵌套仓库开关 #45

+++ git-checkout-core/src/main/kotlin/com/tencent/bk/devops/git/core/service/handler/GitSubmodulesHandler.kt
@@ -81,16 +81,18 @@ class GitSubmodulesHandler(
             command = "git reset --hard",
             recursive = nestedSubmodules
         )
-        val builder = StringBuilder("git clean -fd ")
-        if (enableGitCleanIgnore == true) {
-            builder.append(" -x ")
+        if (enableGitClean) {
+            val builder = StringBuilder("git clean -fd ")
+            if (enableGitCleanIgnore == true) {
+                builder.append(" -x ")
+            }
+            if (enableGitCleanNested == true) {
+                builder.append(" -f ")
+            }
+            git.submoduleForeach(
+                command = builder.toString(),
+                recursive = nestedSubmodules
+            )
         }
-        if (enableGitCleanNested == true) {
-            builder.append(" -f ")
-        }
-        git.submoduleForeach(
-            command = builder.toString(),
-            recursive = nestedSubmodules
-        )
     }
 }
